Simplified github fetcher file
